+++ test/herramientas/HachaMetalTest.java
@@ -12,7 +12,7 @@ import materiales.Piedra;
 
 class HachaMetalTest {
 
-	int durabilidadEsperada = 400 - 10 / 2; // para inicial = 400 y DesgastadorLineal Fuerza = 10 (usa 10/2)
+	float durabilidadEsperada = 400 - 10 / 2; // para inicial = 400 y DesgastadorLineal Fuerza = 10 (usa 10/2)
 
 	@Test
 	public void HachaDeMetalGolpeaMaderaYSeDesgastaElHacha() {
