+++ test/herramientas/HachaTest.java
@@ -14,7 +14,7 @@ public class HachaTest {
     @Test
     void seCreaHachaMaderaConDurabilidadCorrecta(){
         Herramienta hachaMadera = new HachaMadera();
-        int durabilidad = hachaMadera.getDurabilidad();
+        float durabilidad = hachaMadera.getDurabilidad();
 
         assertEquals(durabilidad, 100);
     }
@@ -22,14 +22,14 @@ public class HachaTest {
     @Test
     void seCreaHachaMaderaConFuerzaCorrecta(){
         HachaMadera hachaMadera = new HachaMadera();
-        int fuerza = hachaMadera.getFuerza();
+        float fuerza = hachaMadera.getFuerza();
         assertEquals(fuerza, 2);
     }
 
     @Test
     void seDesgastaUnaHachaDeMadera(){
         HachaMadera hachaMadera = new HachaMadera();
-        int durabilidad = hachaMadera.getDurabilidad();
+        float durabilidad = hachaMadera.getDurabilidad();
         hachaMadera.desgastar();
 
         assertEquals(durabilidad-2, hachaMadera.getDurabilidad());
@@ -38,7 +38,7 @@ public class HachaTest {
     @Test
     void seCreaHachaPiedraConDurabilidadCorrecta(){
         Herramienta hachaPiedra = new HachaPiedra();
-        int durabilidad = hachaPiedra.getDurabilidad();
+        float durabilidad = hachaPiedra.getDurabilidad();
 
         assertEquals(durabilidad, 200);
     }
@@ -46,7 +46,7 @@ public class HachaTest {
     @Test
     void seCreaHachaPiedraConFuerzaCorrecta(){
         Herramienta hachaPiedra = new HachaPiedra();
-        int fuerza = hachaPiedra.getFuerza();
+        float fuerza = hachaPiedra.getFuerza();
 
         assertEquals(fuerza, 5);
     }
@@ -54,7 +54,7 @@ public class HachaTest {
     @Test
     void seDesgastaUnaHachaDePiedra(){
         HachaPiedra hachaPiedra = new HachaPiedra();
-        int durabilidad = hachaPiedra.getDurabilidad();
+        float durabilidad = hachaPiedra.getDurabilidad();
         hachaPiedra.desgastar();
 
         assertEquals(durabilidad-5, hachaPiedra.getDurabilidad());
@@ -63,21 +63,21 @@ public class HachaTest {
     @Test
     void seCreaHachaMetalConDurabilidadCorrecta(){
         HachaMetal hachaMetal= new HachaMetal();
-        int durabilidad = hachaMetal.getDurabilidad();
+        float durabilidad = hachaMetal.getDurabilidad();
         assertEquals(400, durabilidad);
     }
 
     @Test
     void seCreaHachaMetalConFuerzaCorrecta(){
         HachaMetal hachaMetal = new HachaMetal();
-        int fuerza = hachaMetal.getFuerza();
+        float fuerza = hachaMetal.getFuerza();
         assertEquals(10,fuerza);
     }
 
     @Test
     void seCreaHachaMetalSeDesgastaCorrectamente(){
         HachaMetal hachaMetal = new HachaMetal();
-        int durabilidad = hachaMetal.getDurabilidad();
+        float durabilidad = hachaMetal.getDurabilidad();
         hachaMetal.desgastar();
         assertEquals(durabilidad-(hachaMetal.getFuerza()/2),hachaMetal.getDurabilidad());
     }
