+++ test/herramientas/desgastador/DesgastadorLinealTest.java
@@ -10,14 +10,14 @@ public class DesgastadorLinealTest {
     @Test
     void seDesgastaLinealMente(){
         DesgastadorLineal desgatadorLineal = new DesgastadorLineal(1,1);
-        int desgastado = desgatadorLineal.calcularDesgaste();
+        float desgastado = desgatadorLineal.calcularDesgaste();
         assertEquals(desgastado,1);
     }
 
     @Test
     void seDesgastaDiezVecesCorrectamente(){
         DesgastadorLineal desgatadorLineal = new DesgastadorLineal(1,1);
-        int desgastado = 0;
+        float desgastado = 0;
         for (int i=0; i<10; i++) {
             desgastado += desgatadorLineal.calcularDesgaste();
         }
@@ -27,7 +27,7 @@ public class DesgastadorLinealTest {
     @Test
     void seDesgastaLinealMenteConFuerzaCuatroYInDiceDesgasteDos(){
         DesgastadorLineal desgatadorLineal = new DesgastadorLineal(4,2);
-        int desgastado = desgatadorLineal.calcularDesgaste();
+        float desgastado = desgatadorLineal.calcularDesgaste();
         assertEquals(desgastado,2);
     }
