+++ test/materiales/MetalTest.java
@@ -13,7 +13,7 @@ public class MetalTest {
     @Test
     public void metalEsGolpeadoConPicoDeMaderaYNoSeReduceSuDurabilidad() {
     	Material metal = new Metal();
-        int durabilidad = metal.getDurabilidad();
+        float durabilidad = metal.getDurabilidad();
         Herramienta pico = new PicoMadera();
 
         metal.desgastarCon(pico);
@@ -24,7 +24,7 @@ public class MetalTest {
     @Test
     public void metalEsGolpeadoConPicoDePiedraYSeReduceSuDurabilidad() {
     	Material metal = new Metal();
-        int durabilidad = metal.getDurabilidad();
+        float durabilidad = metal.getDurabilidad();
         Herramienta pico = new PicoPiedra();
 
         metal.desgastarCon(pico);
@@ -35,7 +35,7 @@ public class MetalTest {
     @Test
     public void metalEsGolpeadoConPicoDeMetalYSeReduceSuDurabilidad() {
     	Material metal = new Metal();
-        int durabilidad = metal.getDurabilidad();
+        float durabilidad = metal.getDurabilidad();
         Herramienta pico = new PicoMetal();
 
         metal.desgastarCon(pico);
@@ -46,7 +46,7 @@ public class MetalTest {
     @Test
     public void metalEsGolpeadoConPicoFinoaYSeReduceSuDurabilidad() {
     	Material metal = new Metal();
-        int durabilidad = metal.getDurabilidad();
+        float durabilidad = metal.getDurabilidad();
         Herramienta pico = new PicoFino();
 
         metal.desgastarCon(pico);
